Get rid of an unneeded variable.
We don't want to try to suppress warnings about discarding the const
qualifier, as that's something we need to fix.  Adding the extra
variable doesn't suppress it on some platforms, so it's not useful
there, and if it does suppress it, we don't want that.

And we weren't doing it in the other case, anyway.

Comment in both cases with an XXX comment about the cast.